--- advent_of_code/aoc/2022/inputs/day09_sample.txt
@@ -1,8 +1,0 @@
-R 4
-U 4
-L 3
-D 1
-R 4
-D 1
-L 5
-R 2
